requirements.txt
torch
opencv-python
transformers
pillow
python-dotenv
cloudinary
requests